backend/requirements.txt
# Web Framework
fastapi==0.127.0
uvicorn[standard]==0.24.0
starlette==0.50.0
annotated-doc==0.0.4

# Database
sqlalchemy==2.0.45
alembic==1.12.1
psycopg2-binary==2.9.11
greenlet==3.3.0

# Data Processing
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
ydata-profiling==4.6.4

# Visualization
plotly==5.17.0

# Background Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# File Storage
boto3==1.34.0
minio==7.2.0

# Validation and Serialization
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
email-validator==2.3.0
dnspython==2.8.0
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2025.11.12

# Testing
pytest==9.0.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.148.8
colorama==0.4.6
exceptiongroup==1.3.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
sortedcontainers==2.4.0
tomli==2.3.0

# Development
python-dotenv==1.2.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
anyio==4.12.0
idna==3.11

# Monitoring
prometheus-client==0.19.0